refactor componentSideList
